pyproject.toml
[project]
name = "weather-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.2",
    "apache-airflow-providers-docker>=4.4.0",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
